% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocCache.R
\name{biocCache}
\alias{biocCache}
\title{Use the Bioconductor caching function}
\usage{
biocCache(cache, key, save, update = FALSE)
}
\arguments{
\item{cache}{A BiocFileCache object.}

\item{key, save}{See \code{\link{getFileMetadata}} for details on the caching function.}

\item{update}{Logical scalar indicating whether assets in the cache should be forcibly updated.
Useful for resolving corrupted caches or incomplete downloads.}
}
\value{
String containing a path to the cached resource inside \code{cache}.
}
\description{
This implements one possible \code{cache} function for use in \code{\link{getFileMetadata}} and friends.
Caching is based on \pkg{BiocFileCache} with thread-safety enforced by \pkg{filelock}.
}
\examples{
library(BiocFileCache)
tmp <- tempfile()
bfc <- BiocFileCache(tmp, ask=FALSE)
cache.fun <- function(key, save) biocCache(bfc, key, save)

# First query hits the cache:
getFile(example.id, example.url, cache = cache.fun)

# Second query re-uses the cache:
getFile(example.id, example.url, cache = cache.fun)

}
\seealso{
\code{\link{getFileMetadata}} for more details on the caching function.
}
\author{
Aaron Lun
}
