% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloneProjectVersion.R
\name{cloneProjectVersion}
\alias{cloneProjectVersion}
\title{Clone a version of project}
\usage{
cloneProjectVersion(
  dir,
  url,
  project,
  version,
  link.only = FALSE,
  cache = NULL
)
}
\arguments{
\item{dir}{String containing a path to the destination directory.}

\item{url}{String containing the URL to the REST API.}

\item{project}{String containing the project name.}

\item{version}{String containing the project version.}

\item{link.only}{Logical scalar indicating whether to create placeholder links instead of copies of the non-metadata files.}

\item{cache}{Caching function to use in \code{\link{getFile}} or \code{\link{getFileMetadata}}.}
}
\value{
A directory is created at \code{dir} with the contents of the specified project version.
\code{NULL} is invisibly returned.
}
\description{
Clone the contents of a project version onto the local filesystem.
}
\details{
If \code{link.only=TRUE} is specified, all non-metadata files are replaced by placeholder links using \code{\link{createPlaceholderLink}}.
This avoids the download while retaining the project structure for further modifications, e.g., using \code{\link{initializeUpload}}.

If \code{cache} is provided, metadata documents are first cached and then copied to \code{dir}.
This avoids re-downloads of the same documents in subsequent \code{\link{cloneProjectVersion}} calls.
If \code{link.only=FALSE}, non-metadata files will also be cached.
}
\examples{
tmp <- tempfile()
cloneProjectVersion(tmp, example.url, example.project, example.version)
list.files(tmp, recursive=TRUE)

}
\seealso{
\code{\link{createPlaceholderLink}}, for the link creation procedure.
}
\author{
Aaron Lun
}
