% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getChildMetadata.R
\name{getChildMetadata}
\alias{getChildMetadata}
\title{Retrieve metadata for a file and its children}
\usage{
getChildMetadata(id, ...)
}
\arguments{
\item{id}{String containing the ArtifactDB identifier for a file.
This is a concatenated identifier involving the project name, file path and version,
i.e., \code{<project>:<path>@<version>} (see Examples).}

\item{...}{Further arguments to pass to \code{\link{getFileMetadata}}.}
}
\value{
Named list where each entry corresponds to a file and contains the metadata for that file.
The first entry is always the metadata for \code{id}, and the remaining entries contain the metadata for children found in a depth-first search.
}
\description{
Retrieve metadata for a file and all its children from an ArtifactDB instance.
A file's set of children is defined as all other files that are referenced from the first file's metadata via the special \code{resource} property.
}
\examples{
# Not particularly interesting.
X <- getChildMetadata(example.id, url = example.url)
str(X)

}
\author{
Aaron Lun
}
