% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProjectMetadata.R
\name{getProjectMetadata}
\alias{getProjectMetadata}
\title{Retrieve metadata for a project}
\usage{
getProjectMetadata(project, url, version = NULL)
}
\arguments{
\item{project}{String containing the name of a project.}

\item{url}{String containing the URL of the ArtifactDB REST endpoint.}

\item{version}{String specifying the version or revision of the project for which to retrieve metadata.}
}
\value{
A list is returned containing metadata for all files from a project.
If \code{version} is specified, files are only returned for that versino of the project.
}
\description{
Retrieve metadata for a project from an ArtifactDB using its REST endpoints.
}
\examples{
# Project-wide metadata for a single version:
out <- getProjectMetadata(example.project, version=example.version, url=example.url)
length(out)

# Project-wide metadata, all versions:
out <- getProjectMetadata(example.project, url = example.url)
length(out)

}
\seealso{
\code{\link{getFileMetadata}}, to get metadata for a specific file.
}
\author{
Aaron Lun
}
