% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listProjects.R
\name{listProjects}
\alias{listProjects}
\title{List projects}
\usage{
listProjects(url, number = 50)
}
\arguments{
\item{url}{String containing the URL to the ArtifactDB REST API.}

\item{number}{Number of projects to report.}
}
\value{
Named list of character vectors.
Each character vector corresponds to a project and is named after that project;
the vector itself contains the names of all versions of that project.
The list is of length no greater than \code{number}.
}
\description{
List the available projects that are visible to the current user.
}
\examples{
listProjects(example.url, number = 10)

}
\seealso{
\code{\link{listProjectVersions}}, to obtain a listing of all projects.
}
\author{
Aaron Lun
}
