% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolveLatest.R
\name{resolveLatestVersion}
\alias{resolveLatestVersion}
\alias{resolveLatestID}
\title{Resolve the latest version}
\usage{
resolveLatestVersion(project, url)

resolveLatestID(id, url)
}
\arguments{
\item{project}{String containing the project identifier.}

\item{url}{String containing the URL of the ArtifactDB endpoint.}

\item{id}{String containing a file identifier.}
}
\value{
For \code{resolveLatestVersion}, a string containing the latest version in its canonical form.

For \code{resolveLatestID}, \code{id} is updated with any \code{"latest"} version alias replaced by its canonical form.
}
\description{
Figure out the latest version of a project or file.
}
\examples{
resolveLatestVersion(example.project, example.url)

unpacked <- unpackID(example.id)
unpacked$version <- "latest"
latest.id <- do.call(packID, unpacked)
resolveLatestID(latest.id, example.url)

}
\author{
Aaron Lun
}
