% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchMetadata.R
\name{searchMetadata}
\alias{searchMetadata}
\title{Search on metadata}
\usage{
searchMetadata(
  q,
  url,
  fields = NULL,
  number = NULL,
  latest = TRUE,
  warn.auth = TRUE,
  transient = FALSE,
  quote.dash = TRUE,
  user.agent = NULL
)
}
\arguments{
\item{q}{String containing an Elasticsearch-compatible query.}

\item{url}{String containing the URL of the ArtifactDB REST endpoint.}

\item{fields}{Character vector containing the fields to return in the metadata for each matching file.
If \code{NULL}, all fields are returned.}

\item{number}{Integer scalar specifying the maximum number of records to return.
Defaults to 50.}

\item{latest}{Logical scalar indicating whether to only return files for the latest revision of each project.}

\item{warn.auth}{Logical scalar indicating whether to warn the user if they are not authenticated.}

\item{transient}{Logical scalar indicating whether to return transient files, i.e., with an expiry date.}

\item{quote.dash}{Logical scalar indicating whether words containing dashes should quoted to avoid tokenization of the components.}

\item{user.agent}{String specifying the user agent, defaults to the current R installation and \pkg{zircon} version.}
}
\value{
A list of lists where each inner list corresponds to a matching file and contains its metadata.
The metadata entry for each file is equivalent to that returned by \code{\link{getFileMetadata}}.
}
\description{
Search an ArtfiactDB based on the file/project metadata.
}
\details{
If \code{latest=TRUE} and \code{fields} is not \code{NULL}, the output list will contain extra fields required for sorting versions.
These have no effect on the other returned values and can be safely ignored if they are not of interest.

By default, if the user is not authenticated, the function will only return public results with a warning.
The warning can be disabled with \code{warn=FALSE} if the public restriction is intentional.c

If \code{number=NULL}, a warning is raised if the search results are truncated.
}
\examples{
searchMetadata(example.project, field=c("path", "description"), url=example.url)

}
\author{
Aaron Lun
}
