% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpackID.R
\name{unpackID}
\alias{unpackID}
\alias{packID}
\title{Pack or unpack a SebasDB ID}
\usage{
unpackID(id)

packID(project, path, version)
}
\arguments{
\item{id}{String containing an identifier for a file.}

\item{project}{String containing the project name. 
This should not contain \code{:}.}

\item{path}{String containing the relative rooted path to the result within the analysis project.}

\item{version}{String containing the project version.
This should not contain \code{@}.}
}
\value{
\code{unpackID} will break down \code{id} and return a list containing character vectors of its components.
Note that the \code{"commit"} field is deprecated and is only returned for backwards compatibility.

\code{packID} will pack the components into a single string.
If parallel components are supplied, a character vector is returned.
}
\description{
Pack or unpack an identifier for a file in SebasDB.
}
\examples{
unpackID(example.id)
packID(example.project, "blah.rds", example.version)

}
\author{
Aaron Lun
}
