% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{zircon-examples}
\alias{zircon-examples}
\alias{example.url}
\alias{example.id}
\alias{example.project}
\alias{example.version}
\title{Example values for testing}
\description{
Example instances of the ArtifactDB API and its data, for use in \pkg{zircon}'s own unit tests.
}
