% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{setGitHubToken}
\alias{setGitHubToken}
\alias{getGitHubTokenInfo}
\alias{getJWTFromGitHub}
\alias{useGitHubIdentities}
\title{Authenticating via GitHub}
\usage{
setGitHubToken(
  token,
  cache.env,
  cache.path = NULL,
  github.url = "https://api.github.com",
  user.agent = NULL,
  prompt.text = NULL
)

getGitHubTokenInfo(cache.env, cache.path = NULL, prompt = interactive(), ...)

getJWTFromGitHub(
  cache.env,
  client.id,
  jwt.cache.path = NULL,
  gh.cache.path = NULL,
  prompt = interactive(),
  user.agent = NULL,
  ...
)

useGitHubIdentities(
  cache.env,
  jwt.cache.path = NULL,
  gh.cache.path = NULL,
  ...
)
}
\arguments{
\item{token}{String containing a GitHub personal access token.
If missing, the user will be prompted to supply a token.
If \code{NULL}, any existing tokens are cleared.}

\item{cache.env}{Environment to use to store the current GitHub token in an in-memory cache.
This allows the same token to be transparently re-used in the same R session.}

\item{cache.path}{String containing a path to a on-disk cache.
This allows the same token to be transparently re-used across different R sessions.
If \code{NULL}, no on-disk caching is performed.}

\item{github.url}{String containing the URL to a GitHub REST API.}

\item{user.agent}{String specifying the user agent for queries to various endpoints.}

\item{prompt.text}{String containing the prompt text for an interactively supplied token.}

\item{prompt}{Logical scalar indicating whether the user should be prompted to supply token details if no cached token exists.}

\item{...}{Further arguments to pass to \code{setGitHubToken} when prompting for a new token.}
}
\value{
For \code{setGitHubToken}, any non-\code{NULL} \code{token} (or its interactively supplied counterpart) will be cached in memory and/or on disk, depending on \code{cache.path}.
A list is invisibly returned containing details about the token including its value, the corresponding user account and the expiry time.
If \code{token=NULL}, any cached token is cleared from disk and memory, and \code{NULL} is invisibly returned.

For \code{getGitHubTokenInfo}, a list is returned containing the token details.
If \code{prompt=FALSE} and no cached token is present, \code{NULL} is returned.

For \code{useGitHubIdentities}, the availability and value of the token is used to configure \code{\link{identityAvailable}} and \code{\link{identityHeaders}}.
A function is returned that, when executed, restores the previous values of \code{\link{identityAvailable}} and \code{\link{identityHeaders}};
this is typically run in an \code{\link{on.exit}} clause.
}
\description{
Configure \code{\link{authorizedVerb}} to use GitHub personal access tokens for authentication.
This is a useful default approach when no better authentication scheme is available.
}
\examples{
\dontrun{setGitHubToken()}

cache.env <- new.env()
cache.path <- tempfile()
getGitHubTokenInfo(cache.env, cache.path, prompt=FALSE)

}
\author{
Aaron Lun
}
