% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploadProject.R
\name{uploadProject}
\alias{uploadProject}
\title{Upload a new version of a project}
\usage{
uploadProject(
  dir,
  url,
  project,
  version,
  files = list.files(dir, recursive = TRUE),
  ...,
  permissions = list(),
  upload.args = list(),
  complete.args = list(),
  user.agent = NULL
)
}
\arguments{
\item{dir}{String containing the path to a project directory on the file system, containing files to be uploaded.}

\item{url}{String containing the URL to the REST API.}

\item{project}{String containing the project name.}

\item{version}{String containing the version.}

\item{files}{Character vector of paths to files to be uploaded within \code{dir}.
These should be relative to \code{dir}.}

\item{...}{Further arguments to pass to \code{\link{initializeUpload}}.}

\item{permissions}{A list containing permission information, see \code{\link{getPermissions}}.
This usually contains the \code{owners} and \code{viewers} character vectors.}

\item{upload.args}{Further arguments to pass to \code{\link{uploadFiles}}.}

\item{complete.args}{Further arguments to pass to \code{\link{completeUpload}}.}

\item{user.agent}{String containing a user agent string.
If \code{NULL}, a default user agent is used.}
}
\value{
On success, \code{NULL} is returned invisibly.
}
\description{
Upload a new version of a project, given a staging directory containing files and their metadata.
}
\details{
Use of this function will almost always require appropriate authentication/authorization with the target API.
Developers should ensure that \code{\link{identityHeaders}} and friends are set accordingly.

Users should consider setting \code{expires} in \code{...} when testing uploads, to avoid creating a permanent copy of test files.
Check out \code{?\link{initializeUpload}} for more details.

Advanced users can reduce upload bandwidth and storage footprint by creating links to duplicate files between versions or projects.
Check out \code{?\link{initializeUpload}} for more details.
}
\examples{
# Creating a mock project.
src <- system.file("scripts", "mock.R", package="zircon")
source(src)
tmp <- tempfile()
createMockProject(tmp)

# Performing a basic upload:
\dontrun{
uploadProject(tmp, example.url, "test-zircon-upload", "foobar", expires=1)
}
}
\seealso{
\code{\link{uploadFiles}} and related functions, for the internal utilities.
}
\author{
Aaron Lun
}
